+++ src/max/mod.rs
@@ -1,5 +1,7 @@
 //! Maximize harness values.
 
 pub use harness::MaxHarness;
+pub use value::Value;
 
 pub mod harness;
+pub mod value;